mv inline.article to InlineMixin.article
