+++ botogram/inline.py
@@ -17,30 +17,6 @@
 #   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 #   FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 #   DEALINGS IN THE SOFTWARE.
-from . import syntaxes
-
-
-class Inline:
-
-    @staticmethod
-    def article(title, text, syntax=None, preview=True, attach=None):
-        args = {"type": "article",
-                "id": None,
-                "title": title,
-                "input_message_content": {
-                    "message_text": text,
-                    "disable_web_page_preview": not preview
-                }
-                }
-        if attach is not None:
-            if not hasattr(attach, "_serialize_attachment"):
-                raise ValueError("%s is not an attachment" % attach)
-            args["reply_markup"] = \
-                attach._serialize_attachment("-100200000")
-        syntax = syntaxes.guess_syntax(text, syntax)
-        if syntax is not None:
-            args["input_message_content"]["parse_mode"] = syntax
-        return args
 
 
 def process(bot, chains, update):
